Move lowerTryExprs & support to its own file
and also adjust a comment
